### 2025-03-22 3b46b76
changed the readme for a better way that users can install the package and use it
