### 2025-04-02 #7936 @jbytecode
@crvernon - Do we need a scope check for this submission because of the loc?
