### 2025-05-13 #8199 @editorialbot
[REVIEW]: LMDiskANN.jl: An Implementation of the Low Memory Disk Approximate Nearest Neighbors Search Algorithm
**Submitting author:** @mantzaris (<a href="http://orcid.org/0000-0002-0026-5725">Alexander V. Mantzaris</a>)
**Repository:** https://github.com/mantzaris/LMDiskANN.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
