### 2025-05-18 mantzaris/LMDiskANN.jl#1 @findmyway
About the naming convention in API
Both `camelStyle` (like `createIndex`) and `snake_style` are used in the exported APIs. I think it's better to keep them consistent.
