### 2025-05-18 mantzaris/LMDiskANN.jl#2 @findmyway
Add code for reproducing results from the original paper?
The quick start example code is very useful. But I think the potential users of this package are more interested in real world use cases.

I believe adding code to reproduce experiment results from the original paper will help people gain a better understanding of the advantages provided by this package.

I'm not sure whether we can find the Python implementation of Vamana or not. It would be great if we can also add it (or the Julia version) as a baseline for comparison.
