### 2025-05-18 mantzaris/LMDiskANN.jl#3 @findmyway
Some minor suggestions
1. The title format in the reference is better to be aligned with other sections.
2. In README, it uses `julia@v1.9` as example but the compact of `julia` is set to `1.6`. Better to set them as the same?
3. The `LevelDB2` part can be added as an [extension](https://docs.julialang.org/en/v1/manual/code-loading/#man-extensions) to this package?
4. `Euclidean` is hardcoded in code. Can you support setting it by users? (Also `DEFAULT_MAX_DEGREE` and `DEFAULT_MAX_DEGREE`)
5. The design of `LMDiskANNIndex` is a little bit strange to me. `vec` `adj` and `meta` fields point to file paths instead of concrete data structures. (Not a big problem though.)
    - I personally would prefer reading from/saving to a directory instead of a common file prefix.
6. In the `loadIndex` function, user still needs to specify the element type of `T`. Generally this is not a good design. Because a user of a prebuilt index may have no idea of how to set it beforehand. Can we put it into the `meta` file?
