### 2025-05-19 mantzaris/LMDiskANN.jl#1 @mantzaris
As pointed out `createIndex` is not consistent with the majority of snake case usage as well as `loadIndex`, `resizeIndex`, `saveIndex`. These have all now been renamed to `create_index`,  `load_index`, `resize_index`, `save_index`. Not all of them were exported API functions but the general consistency is a good idea.
