### 2025-05-19 090a8be
function name changes to be consistent and the readme
