+++ src/LMDiskANN.jl
@@ -12,7 +12,7 @@ include("UserIdMapping.jl")
 export open_databases, close_databases, insert_key!, get_id_from_key, get_key_from_id
 export delete_by_key!, delete_by_id!, count_entries, clear_database!, clear_all_databases!, list_all_keys
 
-export createIndex, loadIndex, close_id_mapping, ann_insert!, ann_delete!, search
+export create_index, load_index, close_id_mapping, ann_insert!, ann_delete!, search
 export get_embedding_from_id, get_embedding_from_key
 
 const DEFAULT_MAX_DEGREE = 64  # max number of neighbors
@@ -165,14 +165,14 @@ function _init_files(vecfile::String,
     open(adjfile, "w") do f end
 end
 
-"""
-    _compute_distance(x, y)
+# """
+#     _compute_distance(x, y)
 
-Compute Euclidean distance for T=Float32 (the default param).
-"""
-@inline function _compute_distance(x::AbstractVector{Float32}, y::AbstractVector{Float32})
-    return evaluate(Euclidean(), x, y)
-end
+# Compute Euclidean distance for T=Float32 (the default param).
+# """
+# @inline function _compute_distance(x::AbstractVector{Float32}, y::AbstractVector{Float32})
+#     return evaluate(Euclidean(), x, y)
+# end
 
 
 """
@@ -212,7 +212,7 @@ end
 
 
 """
-    createIndex(path_prefix::String, dim::Int; maxdegree::Int=DEFAULT_MAX_DEGREE)
+    create_index(path_prefix::String, dim::Int; maxdegree::Int=DEFAULT_MAX_DEGREE)
 
 Creates a brand new LM-DiskANN index on disk with the given dimension, storing
 to files: `path_prefix.vec`, `path_prefix.adj`, `path_prefix.meta`.
@@ -230,10 +230,10 @@ to files: `path_prefix.vec`, `path_prefix.adj`, `path_prefix.meta`.
 
 # Example
 ```julia
-index = LMDiskANN.createIndex("my_index", 128)
+index = LMDiskANN.create_index("my_index", 128)
 ```
 """
-function createIndex(path_prefix::String, dim::Int;
+function create_index(path_prefix::String, dim::Int;
                      T::Type=Float32,
                      maxdegree::Int=DEFAULT_MAX_DEGREE)
     vecfile  = path_prefix * ".vec"
@@ -259,10 +259,10 @@ function createIndex(path_prefix::String, dim::Int;
 end
 
 """
-    loadIndex(path_prefix::String; T=Float32)
+    load_index(path_prefix::String; T=Float32)
 Loads an existing index, specifying T if it wasn't Float32 originally.
 """
-function loadIndex(path_prefix::String; T::Type=Float32)
+function load_index(path_prefix::String; T::Type=Float32)
     vecfile  = path_prefix * ".vec"
     adjfile  = path_prefix * ".adj"
     metafile = path_prefix * ".meta"
@@ -292,12 +292,12 @@ end
 
 
 """
-    resizeIndex!(index::LMDiskANNIndex{T}, new_size::Int) where {T<:AbstractFloat}
+    resize_index!(index::LMDiskANNIndex{T}, new_size::Int) where {T<:AbstractFloat}
 
 Resizes the memory-mapped files to accommodate at least new_size points.
 This is used internally when the index needs to grow.
 """
-function resizeIndex!(index::LMDiskANNIndex{T}, new_size::Int) where {T<:AbstractFloat}
+function resize_index!(index::LMDiskANNIndex{T}, new_size::Int) where {T<:AbstractFloat}
     vecfile = index.vecfile
     adjfile = index.adjfile
     dim     = index.dim
@@ -317,7 +317,7 @@ end
 
 
[…]
