+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"","page":"Home","title":"Home","text":"CurrentModule = LMDiskANN","category":"page"},{"location":"#LMDiskANN","page":"Home","title":"LMDiskANN","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"LM-DiskANN is an approximate nearest‐neighbor indexing library that builds and maintains a disk‐resident graph structure for high‐dimensional data. It aims to provide a balance between fast search times, memory efficiency, and support for incremental updates (insertion and deletion). Under the hood, LM-DiskANN stores vector embeddings in memory‐mapped files, limiting in‐memory usage while preserving quick random access. It also leverages a graph-based BFS expansion to find candidate neighbors, with user‐configurable parameters like search expansion factor (EF_SEARCH) and maximum neighborhood size (maxdegree).","category":"page"},{"location":"","page":"Home","title":"Home","text":"Pan, Y., Sun, J., & Yu, H. (2023, December). Lm-diskann: Low memory footprint in disk-native dynamic graph-based ann indexing. In 2023 IEEE International Conference on Big Data (BigData) (pp. 5987-5996). IEEE.","category":"page"},{"location":"","page":"Home","title":"Home","text":"It offers the optional user‐key mapping, allowing users to associate string keys with each embedding. This means you can insert vectors with a custom ID (e.g., \"image001\"), search for nearest neighbors, then retrieve or delete by either the numeric internal ID or the user key. The library includes utility functions to open, close, and clear these user‐key databases, enabling a more flexible integration with real-world applications where data might be identified by strings or external IDs.","category":"page"},{"location":"#Example-Usage-with-Synthetic-Data","page":"Home","title":"Example Usage with Synthetic Data","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"Below is a short script you can run to see LM-DiskANN in action. We’ll:","category":"page"},{"location":"","page":"Home","title":"Home","text":"Create an index for dim=5 vectors.\nInsert a few vectors—some with user keys, some without.\nSearch for a known vector.\nDelete by either numeric ID or user key.","category":"page"},{"location":"","page":"Home","title":"Home","text":"using LMDiskANN  \nusing Random\n\n# 1 make an index with dimension 5\ndim = 5\nindex = createIndex(\"example_index\", dim)\n\n# 2 insert a couple of random vectors\nvec1 = rand(Float32, 5)\n(key1, id1) = ann_insert!(index, vec1)\n@show key1, id1   # e.g. might be (\"1\", 1) if no user key was specified\n\n# insert a vector with a user key\nvec2 = rand(Float32, 5)\n(user_key, id2) = ann_insert!(index, vec2; key=\"my_special_vector\")\n@show user_key, id2  # e.g. (\"my_special_vector\", 2)\n\n# 3 search for vec1\nresults = search(index, vec1, topk=3)  # returns an array of (maybeKey, ID)\nprintln(\"Search results for vec1 = \", results)\n\n# 4 retrieve the actual embedding from the index by numeric ID\nretrieved_vec1 = get_embedding_from_id(index, id1)\nprintln(\"retrieved_vec1 == vec1 ? 
[…]
