### 2025-05-19 c0fa0e4
now the user can select the distance function
