### 2025-05-19 mantzaris/LMDiskANN.jl#3 @mantzaris
1. I believe I have reformatted the reference title format to be better and consistent.

2. Yes you are right, this is corrected.

3. The LevelDB2 could be considered an optional extension, but the choice to require it is so that the user can have the string to id mappings 'out of the box' rather than consider having to install another extension. It is a common user request from packages that don't offer it. A common one seen is that they want the embeddings of the files to be mapped to the hash of the files for the lookups.

4.  Distances are now chosen but a default is provided, the tests pass where the user can select eg cosine:

```

@testset "Metric Parameterisation" begin
    clean_up()

    # Euclidean test
    @testset "Default metric is Euclidean" begin
        idx_default = create_index("temp_metric_default", 4)
        @test idx_default.metric isa Euclidean
    end

    # cosine test
    @testset "Cosine distance end-to-end" begin
        dim      = 32
        nvecs    = 500
        topk     = 10
        nqueries = 20
        cosine   = CosineDist()

        idx_cos = create_index("temp_metric_cosine", dim; metric = cosine)

        vecs = [rand(Float32, dim) for _ in 1:nvecs]
        foreach(v -> ann_insert!(idx_cos, v), vecs)

        @test idx_cos.num_points == nvecs
        @test idx_cos.metric === cosine

        # recall check
        total_recall = 0.0
        for q in rand(vecs, nqueries)
            bf = sortperm([evaluate(cosine, q, v) for v in vecs])[1:topk]
            ann_ids = [r[2] for r in search(idx_cos, q, topk = topk)]
            total_recall += length(intersect(Set(bf), Set(ann_ids))) / topk
        end
        avg = total_recall / nqueries
        @info "Average cosine recall = $(round(avg, digits = 3))"
        @test avg ≥ 0.70
        #

        # flush and close so clean_up() can delete the files
        LMDiskANN.save_index(idx_cos)
        close_id_mapping(idx_cos)
    end

    clean_up()
end
```

for the `DEFAULT_MAX_DEGREE`, although defined as a constant, can be selected:

```

function _init_files(vecfile::String,
                     adjfile::String,
                     metafile::String;
                     dim::Int,
                     maxdegree::Int=DEFAULT_MAX_DEGREE)
```

and

```
function create_index(path_prefix::String, dim::Int;
                     T::Type=Float32,
                     maxdegree::Int=DEFAULT_MAX_DEGREE,
                     metric::PreMetric=Euclidean())
```

which is the only places it is used.

5. The choice for a file based approach is to fit into an existing project as simply as possible. It is a file based store essentially and can act like sqlite in that fashion. Users can make a directory if they want.

6. That is a good observation. I have changed that the tests pass and now tests for this show the configuration of types passing:

```
@testset "Element-type configuration" begin
    clean_up()

    for (tag, T) in (("32", Float32), ("16", Float16), ("64", Float64))
        prefix = "temp_tidx" * tag

        # create index with appropriate element type
        idx = (T === Float32) ?
              create_index(prefix, 4) :
              create_index(prefix, 4; T = T)

        @test eltype(idx.vecs) == T
        ann_insert!(idx, rand(T, 4))

        # flush & close so LevelDB directories are written
        LMDiskANN.save_index(idx)
        LMDiskANN.close_id_mapping(idx)

        # reload – element type should be detected from the meta file
        idx2 = load_index(prefix)
        @test eltype(idx2.vecs) == T
    end

    clean_up()
end
```
