### 2025-05-19 fa8fe14
now new ability for the user to set the types of the precision on the vector element data and keep it persist on disk
