+++ src/LMDiskANN.jl
@@ -64,18 +64,18 @@ end
 
 """
     _read_metadata(metafile::String)
-Reads metadata (num_points, dim, maxdegree, freelist, entrypoint) via Serialization.
+Reads metadata (num_points, dim, maxdegree, freelist, entrypoint, eltype) via Serialization.
 """
 function _read_metadata(metafile::String)
     open(metafile, "r") do io
         md = deserialize(io)
-        return md["num_points"], md["dim"], md["maxdegree"], md["freelist"], md["entrypoint"]
+        return md["num_points"], md["dim"], md["maxdegree"], md["freelist"], md["entrypoint"], get(md, "eltype", "Float32")
     end
 end
 
 
 """
-    _write_metadata(metafile::String, num_points, dim, maxdegree, freelist, entrypoint)
+    _write_metadata(metafile::String, num_points::Int, dim::Int, maxdegree::Int, freelist::Vector{Int}, entrypoint::Int, T::Type)
 
 Writes metadata to a `.meta` file using Julia's Serialization.
 """
@@ -84,13 +84,15 @@ function _write_metadata(metafile::String,
                          dim::Int,
                          maxdegree::Int,
                          freelist::Vector{Int},
-                         entrypoint::Int)
+                         entrypoint::Int,
+                         T::Type = Float32)
     md = Dict(
         "num_points" => num_points,
         "dim"        => dim,
         "maxdegree"  => maxdegree,
         "freelist"   => freelist,
-        "entrypoint" => entrypoint
+        "entrypoint" => entrypoint,
+        "eltype"     => string(T)
     )
     open(metafile, "w") do io
         serialize(io, md)
@@ -152,7 +154,7 @@ end
 
 
 """
-    _init_files(vecfile, adjfile, metafile, dim)
+    _init_files(vecfile, adjfile, metafile, dim, maxdegree, T)
 
 Initializes (or overwrites) new files for a fresh index with no points.
 """
@@ -160,12 +162,13 @@ function _init_files(vecfile::String,
                      adjfile::String,
                      metafile::String;
                      dim::Int,
-                     maxdegree::Int=DEFAULT_MAX_DEGREE)
+                     maxdegree::Int=DEFAULT_MAX_DEGREE,
+                     T::Type = Float32)
     num_points = 0
     freelist   = Int[]
     entrypoint = -1
 
-    _write_metadata(metafile, num_points, dim, maxdegree, freelist, entrypoint)
+    _write_metadata(metafile, num_points, dim, maxdegree, freelist, entrypoint, T)
     
     open(vecfile, "w") do f end
     open(adjfile, "w") do f end
@@ -211,7 +214,7 @@ end
 
 
 """
-    create_index(path_prefix::String, dim::Int; maxdegree::Int=DEFAULT_MAX_DEGREE, metric::PreMetric=Euclidean())
+    create_index(path_prefix::String, dim::Int; T::Type=Float32, maxdegree::Int=DEFAULT_MAX_DEGREE, metric::PreMetric=Euclidean())
 
 Creates a brand new LM-DiskANN index on disk with the given dimension, storing
 to files: `path_prefix.vec`, `path_prefix.adj`, `path_prefix.meta`.
@@ -241,8 +244,8 @@ function create_index(path_prefix::String, dim::Int;
     adjfile  = path_prefix * ".adj"
     metafile = path_prefix * ".meta"
 
-    _init_files(vecfile, adjfile, metafile; dim=dim, maxdegree=maxdegree)
-    num_points, dim_, maxdegree_, freelist, entrypoint = _read_metadata(metafile)
+    _init_files(vecfile, adjfile, metafile; dim=dim, maxdegree=maxdegree, T=T)
[…]
