### 2025-05-19 #8199 @mantzaris
@findmyway thanks for the in depth comments and suggestions. There were some points I definitely needed to address to make the package better.

>     * [About the naming convention in API mantzaris/LMDiskANN.jl#1](https://github.com/mantzaris/LMDiskANN.jl/issues/1)

I have addressed these issues, it was great that you picked up on them.

>     * [Add code for reproducing results from the original paper? mantzaris/LMDiskANN.jl#2](https://github.com/mantzaris/LMDiskANN.jl/issues/2)

I have added a more extensive example into the docs, using GLoVe vectors (10,000 of them) which is feasible for most users to attempt and more illustrative for real case scenarios. The datasets from the original paper will probably not be run by most users. Do you disagree with this chosen example?

>     * [Some minor suggestions mantzaris/LMDiskANN.jl#3](https://github.com/mantzaris/LMDiskANN.jl/issues/3)

 I addressed all the issues and it has definitely made the package better having done so.
