+++ Project.toml
@@ -1,7 +1,7 @@
 name = "LMDiskANN"
 uuid = "a2e5363b-c4e1-4593-9e2f-a2937b345987"
 authors = ["Alexander V. Mantzaris"]
-version = "1.1.0-DEV"
+version = "1.1.0"
 
 [deps]
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
