### 2025-05-27 mantzaris/LMDiskANN.jl#2 @findmyway
The `GLoVe` example is very intuitive!

> The example will stay true to the software usability and documentation, not on re-doing the research paper which is obscure for most users and the datasets are too big for most users to even attempt.

I understand that, but as the package owner, trying to reproduce the original results and report them to users will help them make decisions.

---

```jl
using LMDiskANN
using ProgressMeter

function read_vecs(filename::String)
    T = endswith(filename, ".fvecs") ? Float32 : Int32
    data = Vector{T}()
    let dim
        open(filename, "r") do file
            while !eof(file)
                dim_bytes = read(file, 4)
                dim = reinterpret(Int32, dim_bytes)[]
                vector_bytes = read(file, 4 * dim)
                vector = reinterpret(T, vector_bytes)
                append!(data, vector)
            end
        end
        reshape(data, Int(dim), :)
    end
end

function read_dataset(prefix)
    base = read_vecs("$(prefix)_base.fvecs")
    query = read_vecs("$(prefix)_query.fvecs")
    groundtruth = read_vecs("$(prefix)_groundtruth.ivecs")
    return base, query, groundtruth
end

function insertion_latency(idx_name, base)
    stats = [0.0]
    idx = create_index(idx_name, size(base, 1))

    @showprogress for (i, c) in enumerate(eachcol(base))
        t = @elapsed ann_insert!(idx, c)
        if i % (size(base, 2) ÷ 100) == 0
            @info "Stats" i t
            push!(stats, t)
        end
    end
    idx, stats
end

function recall(idx, query, groundtruth, k=10, L=200)
    res = []
    for (v, gt) in zip(eachcol(query), eachcol(groundtruth))
        candidates = search(idx, v, topk=L)
        c = [i - 1 for (k, i) in candidates]
        push!(res, length(intersect(c, gt[1:k])) / k)
    end
    return res
end
```

I used the above code for a quick test on the sift1M dataset. It takes about 2h30min on my local machine to build the whole index. However, it is reported to take `2467s` in the original paper. This is a huge difference.

```julia
julia> versioninfo()
Julia Version 1.11.5
Commit 760b2e5b739 (2025-04-14 06:53 UTC)
Build Info:
  Official https://julialang.org/ release
Platform Info:
  OS: macOS (arm64-apple-darwin24.0.0)
  CPU: 16 × Apple M4 Max
  WORD_SIZE: 64
  LLVM: libLLVM-16.0.6 (ORCJIT, apple-m1)
Threads: 12 default, 0 interactive, 6 GC (on 12 virtual cores)
```

And the latency is quite different compared to the original paper

![Image](https://github.com/user-attachments/assets/16a19f3c-7179-42df-88e2-b1b8fcec5e2b)

The x-axis is the percentage of inserted embeddings and the y axis is latency sampling (in seconds).

For comparision:

Another strange point is, I find that the `10-recall@10` is too high even with a small candidate size `L`.

```
julia> recall(idx, sift[2], sift[3], 10, 10) |> mean
0.9936299999999993

julia> recall(idx, sift[2], sift[3], 10, 100) |> mean
0.9941999999999994

julia> recall(idx, sift[2], sift[3], 10, 200) |> mean
0.9941999999999994
```

After examining the code, I notice that in the following line:

https://github.com/mantzaris/LMDiskANN.jl/blob/a9d824079bffe324ffc18cdf39c734cd185564b2/src/LMDiskANN.jl#L461

`topk` is set to a lower bound of `EF_SEARCH`, maybe we can set `EF_SEARCH` as a default value instead of cliping `topk` here?
