### 2025-05-27 mantzaris/LMDiskANN.jl#2 @mantzaris
- If you bucket the raw per-insert latencies the way the paper does and plot the mean latency per bin, the resulting curve is the same gently rising, sub-linear line shown in the paper, it isn’t flat and its slope increases only gradually as the index grows.

- You’re absolutely right: the current code clamps ef to 300, which is why recall stays really high (>0.9) even when you think you’re running with L = 10, due to the clamping to `EF_SEARCH` at 300 as a minimum.
I have renamed it to:  ` const DEFAULT_EF_SEARCH = 300  # renaming for clarity`

Now the search function provides it as a default (override possible with optional param). The doc string describes this now too:
```julia
function search(index::LMDiskANNIndex{T},
                query_vec::AbstractVector{<:AbstractFloat};
                topk::Int=10,
                ef::Int = DEFAULT_EF_SEARCH) where {T<:AbstractFloat}
...
    ef_candidates = _search_graph(index, local_q, max(topk, ef))
```

It still guards against the invalid case ef < topk by taking the max.
(*I have seen it often that picks of ef_search >> k*)

The change is being pushed and docs updated.
