### 2025-05-27 995b03b
now the docs for search are improved, the ef param is set as a default for the search so the user can optionally set it to be lower than the default trading accuracy for speed.
