+++ paper/paper.bib
@@ -5,14 +5,16 @@
   booktitle={2023 IEEE International Conference on Big Data (BigData)},
   pages={5987--5996},
   year={2023},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/BigData59044.2023.10386517}
 }
 
 @article{singh2021freshdiskann,
   title={Freshdiskann: A fast and accurate graph-based ann index for streaming similarity search},
   author={Singh, Aditi and Subramanya, Suhas Jayaram and Krishnaswamy, Ravishankar and Simhadri, Harsha Vardhan},
   journal={arXiv preprint arXiv:2105.09613},
-  year={2021}
+  year={2021},
+  doi={10.48550/arXiv.2105.09613}
 }
 
 @article{jayaram2019diskann,
@@ -27,9 +29,9 @@
   title={A comprehensive survey and experimental comparison of graph-based approximate nearest neighbor search},
   author={Wang, Mengzhao and Xu, Xiaoliang and Yue, Qiang and Wang, Yuxiang},
   journal={arXiv preprint arXiv:2101.12631},
-  year={2021}
+  year={2021},
+  doi = {10.48550/arXiv.2101.12631}
 }
-
 
 @article{nene1997simple,
   title={A simple algorithm for nearest neighbor search in high dimensions},
@@ -39,5 +41,29 @@
   number={9},
   pages={989--1003},
   year={1997},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/34.615448}
+}
+
+@article{Tellez2022, 
+doi = {10.21105/joss.04442}, 
+url = {https://doi.org/10.21105/joss.04442}, 
+year = {2022}, 
+publisher = {The Open Journal}, 
+volume = {7}, 
+number = {75}, 
+pages = {4442}, 
+author = {Eric S. Tellez and Guillermo Ruiz}, 
+title = {`SimilaritySearch.jl`: Autotuned nearest neighbor indexes for Julia}, 
+journal = {Journal of Open Source Software} 
+} 
+
+@misc{tellez2022similarity,
+      title={Similarity search on neighbor's graphs with automatic Pareto optimal performance and minimum expected quality setups based on hyperparameter optimization}, 
+      author={Eric S. Tellez and Guillermo Ruiz},
+      year={2022},
+      eprint={2201.07917},
+      archivePrefix={arXiv},
+      primaryClass={cs.IR},
+      doi = {10.48550/arXiv.2201.07917}
 }
