### 2025-06-06 #8199 @oxinabox
A minor thing:

>  The package SimilaritySearch.jl Tellez & Ruiz (2022b), offers the most mature codebase

probably should have been:

>  The package SimilaritySearch.jl (Tellez & Ruiz, 2022b), offers the most mature codebase
