### 2025-06-06 #8199 @oxinabox
It would be nice, but not strictly required for there to be experimental data showing **Reduced Memory Overhead** vs some baseline that was not on disk.
Similarly for:

> High Recall: Tuning BFS expansions and adjacency degrees can yield high-quality nearest neighbor results

It would be nice to see on some benchmark dataset a plot of how the recall scales with those parameters.
