### 2025-06-06 #8199 @oxinabox
I note that while the installation does processed as documented, this package has not been registered in the julia General Registry.

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

This seems missing.

Other than that looks good. I like the longer example in the docs.
(Not solely because it uses one of my old packages that i didn't know people still used)
