### 2025-06-06 1393de0
now the community notes are added to the docs and the references to the paper improved
