+++ paper/paper.bib
@@ -45,6 +45,20 @@
   doi = {10.1109/34.615448}
 }
 
+@article{tellez2022similarity,
+      title={Similarity search on neighbor graphs with automatic Pareto optimal performance and minimum expected quality setups based on hyperparameter optimization}, 
+      author={Eric S. Tellez and Guillermo Ruiz},
+      journal = {CoRR},
+      year={2022},
+      volume = {abs/2201.07917},
+      eprint={2201.07917},
+      eprinttype = {arXiv},
+      archivePrefix={arXiv},
+      primaryClass={cs.IR},
+      biburl = {https://dblp.org/rec/journals/corr/abs-2201-07917.bib},
+      doi = {10.48550/arXiv.2201.07917}
+}
+
 @article{Tellez2022, 
 doi = {10.21105/joss.04442}, 
 url = {https://doi.org/10.21105/joss.04442}, 
@@ -54,16 +68,8 @@ volume = {7},
 number = {75}, 
 pages = {4442}, 
 author = {Eric S. Tellez and Guillermo Ruiz}, 
-title = {`SimilaritySearch.jl`: Autotuned nearest neighbor indexes for Julia}, 
+title = {SimilaritySearch.jl: Autotuned nearest neighbor indexes for Julia}, 
 journal = {Journal of Open Source Software} 
-} 
-
-@misc{tellez2022similarity,
-      title={Similarity search on neighbor's graphs with automatic Pareto optimal performance and minimum expected quality setups based on hyperparameter optimization}, 
-      author={Eric S. Tellez and Guillermo Ruiz},
-      year={2022},
-      eprint={2201.07917},
-      archivePrefix={arXiv},
-      primaryClass={cs.IR},
-      doi = {10.48550/arXiv.2201.07917}
 }
+
+
