### 2025-06-10 5968814
add citation to Julia
