+++ v1.1.0-DEV/siteinfo.js
@@ -1,0 +1,2 @@
+var DOCUMENTER_CURRENT_VERSION = "v1.1.0-DEV";
+var DOCUMENTER_IS_DEV_VERSION = false;
