### 2025-06-11 #8199 @mantzaris
@jbytecode , yes, I made a couple of 'tiny' changes to the paper alone, and pushed it.
