### 2025-06-12 #8199 @jbytecode
@mantzaris - The paper and the software repo look now better to me. I'm recommending an acceptance. Our track editor will make the final decision. Thank you again for submitting your work to JOSS.
