### 2025-06-12 #8199 @crvernon
:wave: @mantzaris - This submission is looking great!  Please just address the following minor things before I accept for publication:

In your paper:
- [ ] LINE 43:  All of your other points have an empty line after them; however, there is no empty line in-between 4 and 5.
- [ ] LINE 79:  The "p" in "pareto" should be capitalized.  You can do this in your bib file by placing curly brackets around the characters you wish to maintain formatting for.
- [ ] LINE 83:  The "j" in "julia" should be capitalized.

No new release is required for these changes.  Let me know when they have been addressed and I'll accept.  Thanks!
