### 2025-06-12 975ffd8
improvements for the paper
