+++ paper/paper.bib
@@ -46,7 +46,7 @@
 }
 
 @article{tellez2022similarity,
-      title={Similarity search on neighbor graphs with automatic Pareto optimal performance and minimum expected quality setups based on hyperparameter optimization}, 
+      title={Similarity search on neighbor graphs with automatic {Pareto} optimal performance and minimum expected quality setups based on hyperparameter optimization}, 
       author={Eric S. Tellez and Guillermo Ruiz},
       journal = {CoRR},
       year={2022},
@@ -68,7 +68,7 @@ volume = {7},
 number = {75}, 
 pages = {4442}, 
 author = {Eric S. Tellez and Guillermo Ruiz}, 
-title = {SimilaritySearch.jl: Autotuned nearest neighbor indexes for Julia}, 
+title = {SimilaritySearch.jl: Autotuned nearest neighbor indexes for {Julia}}, 
 journal = {Journal of Open Source Software} 
 }
 
